BGDIINF_SB-1521: Renamed the temporal_extent module to be more consistent
